refactor(auth-pbf): move classes to common package
