+++ backend/src/main/java/com/brokendev/backend/common/domain/account/Account.java
@@ -1,6 +1,7 @@
-package com.brokendev.backend.domain;
+package com.brokendev.backend.common.domain.account;
 
 
+import com.brokendev.backend.common.domain.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
