+++ backend/src/main/java/com/brokendev/backend/common/domain/account/AccountRepository.java
@@ -1,7 +1,6 @@
-package com.brokendev.backend.repositories;
+package com.brokendev.backend.common.domain.account;
 
-import com.brokendev.backend.domain.Account;
-import com.brokendev.backend.domain.User;
+import com.brokendev.backend.common.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
