+++ backend/src/main/java/com/brokendev/backend/services/AccountService.java
@@ -1,9 +1,11 @@
 package com.brokendev.backend.services;
 
-import com.brokendev.backend.domain.Account;
+import com.brokendev.backend.common.domain.account.Account;
+import com.brokendev.backend.common.domain.account.AccountRepository;
+import com.brokendev.backend.common.domain.user.UserRepository;
 import com.brokendev.backend.domain.BoletoPayment;
 import com.brokendev.backend.domain.PixTransaction;
-import com.brokendev.backend.domain.User;
+import com.brokendev.backend.common.domain.user.User;
 import com.brokendev.backend.dto.account.AccountBalanceResponseDTO;
 import com.brokendev.backend.dto.account.AccountDepositResponseDTO;
 import com.brokendev.backend.dto.account.TransactionStatementResponseDTO;
