+++ backend/src/main/java/com/brokendev/backend/services/CardService.java
@@ -1,14 +1,14 @@
 package com.brokendev.backend.services;
 
 
-import com.brokendev.backend.domain.Account;
+import com.brokendev.backend.common.domain.account.Account;
 import com.brokendev.backend.domain.Card;
 import com.brokendev.backend.dto.card.CardBlockResponseDTO;
 import com.brokendev.backend.dto.card.CardCreateRequestDTO;
 import com.brokendev.backend.dto.card.CardResponseDTO;
 import com.brokendev.backend.common.exceptions.AccountNotFoundException;
 import com.brokendev.backend.common.exceptions.CardNotFoundException;
-import com.brokendev.backend.repositories.AccountRepository;
+import com.brokendev.backend.common.domain.account.AccountRepository;
 import com.brokendev.backend.repositories.CardRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
