+++ backend/src/main/java/com/brokendev/backend/services/NotificationService.java
@@ -1,7 +1,7 @@
 package com.brokendev.backend.services;
 
 import com.brokendev.backend.domain.Notification;
-import com.brokendev.backend.domain.User;
+import com.brokendev.backend.common.domain.user.User;
 import com.brokendev.backend.dto.notification.NotificationResponseDTO;
 import com.brokendev.backend.common.exceptions.NotificationAccessDeniedException;
 import com.brokendev.backend.common.exceptions.NotificationNotFoundException;
